pyproject.toml
[project]
name = "kaggle-open-problems-single-cell-perturbations"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Yuya Shintani", email = "yuya.shintani.uplus26e7@gmail.com" }
]
dependencies = [
    "pydantic>=2.5.2",
    "hydra-core>=1.3.2",
    "dacite>=1.8.1",
    "pandas>=2.0.3",
    "pyarrow>=14.0.1",
    "fastparquet>=2023.10.1",
    "scikit-learn>=1.3.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "ruff>=0.1.6",
    "mypy>=1.7.1",
    "ipython>=8.12.3",
    "ipykernel>=6.27.1",
    "pandas-stubs>=2.0.3.230814",
]

[tool.hatch.metadata]
allow-direct-references = true
